Add option for preventing the use of the gsm encoding
If you are unlucky there is a chance some of symbols in your sms get removed.
This happens to chars which are by the GSM 03.38 standard but are
non-ascii (0-127). [I am having problems with the Clickatell
sms gateway. The symbols get removed from time to time depending on the
route taken by the message.]

The current change adds an option which allows you to ignore the use of
the gsm encoding and use only ascii and unicode. By default the gsm
encoding is in use.

To use the option you have to update set it in the ruby code and in
javascript.